# services/api/openapi.yaml
openapi: 3.0.3
info:
  title: tyr.bet API
  version: 1.0.0
  description: "Testnet-only. Session auth = signed `tyr_session` cookie from the passkey routes. Agent routes are signed with the agent key (x-tyr-agent / x-tyr-timestamp / x-tyr-signature, EIP-191 over method, path, timestamp, sha256(body)); paid agent routes answer 402 with an MPP challenge. Live events: GET /ws (see WsEvent)."
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: tyr_session
    agent:
      type: apiKey
      in: header
      name: x-tyr-signature
    admin:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          description: machine-readable reason (e.g. limit, balance, region)
        issues:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required:
              - path
              - message
            additionalProperties: false
          description: request validation failures
      required:
        - error
      additionalProperties: false
      description: Error
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        agentSessionId:
          type: string
          nullable: true
        idempotencyKey:
          type: string
          nullable: true
        hlMarket:
          type: string
          description: outcome id
        side:
          type: string
          enum:
            - "yes"
            - "no"
        isBuy:
          type: boolean
        tif:
          type: string
        size:
          type: string
          description: decimal number as a string (exact, from Postgres numeric)
        price:
          type: string
          description: decimal number as a string (exact, from Postgres numeric)
        filledSize:
          type: string
          description: decimal number as a string (exact, from Postgres numeric)
        avgPx:
          type: string
          nullable: true
          description: decimal number as a string (exact, from Postgres numeric)
        stakeUsd:
          type: string
          nullable: true
          description: decimal number as a string (exact, from Postgres numeric)
        stakeSalt:
          type: string
          nullable: true
        builderFee:
          type: string
          nullable: true
          description: decimal number as a string (exact, from Postgres numeric)
        hlOid:
          type: string
          nullable: true
        simulated:
          type: boolean
          description: paper fill against the live HL testnet book
        execution:
          nullable: true
        status:
          type: string
        step:
          type: string
          description: last completed saga step
        error:
          type: string
          nullable: true
        tempoStakeTx:
          type: string
          nullable: true
        escrowTx:
          type: string
          nullable: true
        solanaOpenTx:
          type: string
          nullable: true
        parentId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        settlement:
          anyOf:
            - $ref: "#/components/schemas/Settlement"
          nullable: true
      required:
        - id
        - userId
        - agentSessionId
        - idempotencyKey
        - hlMarket
        - side
        - isBuy
        - tif
        - size
        - price
        - filledSize
        - avgPx
        - stakeUsd
        - stakeSalt
        - builderFee
        - hlOid
        - simulated
        - status
        - step
        - error
        - tempoStakeTx
        - escrowTx
        - solanaOpenTx
        - parentId
        - createdAt
        - updatedAt
      additionalProperties: false
    Settlement:
      type: object
      properties:
        orderId:
          type: string
        outcome:
          type: string
        pnl:
          type: string
          description: decimal number as a string (exact, from Postgres numeric)
        payoutUsd:
          type: string
          description: decimal number as a string (exact, from Postgres numeric)
        payoutSalt:
          type: string
          nullable: true
        solanaTx:
          type: string
          nullable: true
        escrowPayTx:
          type: string
          nullable: true
        escrowTopUpTx:
          type: string
          nullable: true
        tempoPayoutTx:
          type: string
          nullable: true
        memoHash:
          type: string
          nullable: true
        closeOrderId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - orderId
        - outcome
        - pnl
        - payoutUsd
        - payoutSalt
        - solanaTx
        - escrowPayTx
        - escrowTopUpTx
        - tempoPayoutTx
        - memoHash
        - closeOrderId
        - createdAt
      additionalProperties: false
...
paths:
...
  /api/auth/passkey/register/verify:
    post:
      tags:
        - auth
      description: Verifies the attestation, creates the user, sets the session cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                rawId:
                  type: string
                type:
                  type: string
                  enum:
                    - public-key
                response:
                  type: object
                  properties: {}
                  additionalProperties: true
              required:
                - id
                - rawId
                - type
                - response
              additionalProperties: true
              description: RegistrationResponseJSON / AuthenticationResponseJSON from @simplewebauthn/browser
        description: RegistrationResponseJSON / AuthenticationResponseJSON from @simplewebauthn/browser
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  tempoAddress:
                    type: string
                    pattern: ^0x[0-9a-fA-F]{40}$
                required:
                  - userId
                  - tempoAddress
                additionalProperties: false
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/auth/passkey/login/verify:
    post:
      tags:
        - auth
      description: Verifies the assertion and sets the session cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                rawId:
                  type: string
                type:
                  type: string
                  enum:
                    - public-key
                response:
                  type: object
                  properties: {}
                  additionalProperties: true
              required:
                - id
                - rawId
                - type
                - response
              additionalProperties: true
              description: RegistrationResponseJSON / AuthenticationResponseJSON from @simplewebauthn/browser
        description: RegistrationResponseJSON / AuthenticationResponseJSON from @simplewebauthn/browser
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  tempoAddress:
                    type: string
                    pattern: ^0x[0-9a-fA-F]{40}$
                required:
                  - userId
                  - tempoAddress
                additionalProperties: false
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/me:
    get:
      tags:
        - auth
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  tempoAddress:
                    type: string
                    pattern: ^0x[0-9a-fA-F]{40}$
                  region:
                    type: string
                    nullable: true
                required:
                  - userId
                  - tempoAddress
                  - region
                additionalProperties: false
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/region:
    put:
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                region:
                  type: string
                  description: ISO 3166-1 alpha-2
              required:
                - region
              additionalProperties: false
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  region:
                    type: string
                    nullable: true
                  hedge:
                    anyOf:
                      - type: object
                        properties:
                          eligible:
                            type: boolean
                            enum:
                              - true
                        required:
                          - eligible
                        additionalProperties: false
                      - type: object
                        properties:
                          eligible:
                            type: boolean
                            enum:
                              - false
                          reason:
                            type: string
                        required:
                          - eligible
                          - reason
                        additionalProperties: false
                required:
                  - region
                  - hedge
                additionalProperties: false
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/bets:
    post:
      tags:
        - bets
      description: Place a bet (Flow A). 402 = loss limit exceeded on-chain, 409 = insufficient confidential balance. Replaying an idempotencyKey returns the same order.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                outcome:
                  type: integer
                  minimum: 0
                side:
                  type: string
                  enum:
                    - "yes"
                    - "no"
                stakeUsd:
                  type: number
                  exclusiveMinimum: true
                  minimum: 0
                  maximum: 1000
                idempotencyKey:
                  type: string
                  minLength: 8
                  maxLength: 128
                maxPrice:
                  type: number
                  exclusiveMinimum: true
                  minimum: 0
                  exclusiveMaximum: true
                  maximum: 1
              required:
                - outcome
                - side
                - stakeUsd
                - idempotencyKey
              additionalProperties: false
      security:
        - session: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "402":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags:
        - bets
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  bets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
                required:
                  - bets
                additionalProperties: false
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/bets/{id}:
    get:
      tags:
        - bets
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/agent/bets:
    post:
      tags:
        - agent
      description: "Paid: the MPP charge (= stake) is capped on-chain by the session, then Flow A runs. Replaying an idempotencyKey returns the same order (200) without a new charge."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                outcome:
                  type: integer
                  minimum: 0
                side:
                  type: string
                  enum:
                    - "yes"
                    - "no"
                stakeUsd:
                  type: number
                  exclusiveMinimum: true
                  minimum: 0
                  maximum: 1000
                idempotencyKey:
                  type: string
                  minLength: 8
                  maxLength: 128
                maxPrice:
                  type: number
                  exclusiveMinimum: true
                  minimum: 0
                  exclusiveMaximum: true
                  maximum: 1
              required:
                - outcome
                - side
                - stakeUsd
                - idempotencyKey
              additionalProperties: false
      security:
        - agent: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "402":
          description: "HTTP 402 MPP challenge (WWW-Authenticate: Payment …); pay on Tempo Moderato and retry"
          content:
            application/json:
              schema:
                description: "HTTP 402 MPP challenge (WWW-Authenticate: Payment …); pay on Tempo Moderato and retry"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/agent/bets/{id}:
    get:
      tags:
        - agent
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
      security:
        - agent: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...